backend/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dominate==2.9.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
requests==2.31.0
SQLAlchemy==2.0.25
typing_extensions==4.13.2
urllib3==2.4.0
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.0.1
